Better unit scale for dataset scale (#3886)
* moved format functions from dataset info tab to format utils

* reworked format scale to adapt to best units

* fixed error

* removed useless comment
