+++ frontend/javascripts/libs/format_utils.js
@@ -41,11 +41,40 @@ export function formatTuple(tuple: ?(Array<number> | Vector3 | Vector6)) {
 
 export function formatScale(scaleArr: Vector3, roundTo?: number = 2): string {
   if (scaleArr != null && scaleArr.length > 0) {
-    const scaleArrRounded = scaleArr.map(value => Utils.roundTo(value, roundTo));
-    return `${scaleArrRounded.join(ThinSpace + MultiplicationSymbol + ThinSpace)} nm³`;
+    let unit = "nm³";
+    let scaleArrAdjusted = scaleArr;
+    const smallestValue = Math.min(...scaleArr);
+    if (smallestValue > 1000000) {
+      scaleArrAdjusted = scaleArr.map(value => value / 1000000);
+      unit = "mm³";
+    } else if (smallestValue > 1000) {
+      scaleArrAdjusted = scaleArr.map(value => value / 1000);
+      unit = "μm³";
+    }
+    const scaleArrRounded = scaleArrAdjusted.map(value => Utils.roundTo(value, roundTo));
+    return `${scaleArrRounded.join(ThinSpace + MultiplicationSymbol + ThinSpace)} ${unit}`;
   } else {
     return "";
   }
+}
+
+export function formatNumberToLength(numberInNm: number): string {
+  if (numberInNm < 1000) {
+    return `${numberInNm.toFixed(0)}${ThinSpace}nm`;
+  } else if (numberInNm < 1000000) {
+    return `${(numberInNm / 1000).toFixed(1)}${ThinSpace}μm`;
+  } else {
+    return `${(numberInNm / 1000000).toFixed(1)}${ThinSpace}mm`;
+  }
+}
+
+export function formatExtentWithLength(
+  extent: Object,
+  formattingFunction: number => string,
+): string {
+  return `${formattingFunction(extent.width)}${ThinSpace}×${ThinSpace}${formattingFunction(
+    extent.height,
+  )}${ThinSpace}×${ThinSpace}${formattingFunction(extent.depth)}`;
 }
 
 export function formatMilliseconds(durationMilliSeconds: number): string {
